pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-flow"
version = "0.1.0"
description = "CLI-first local research paper discovery and reading workflow."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "jinja2>=3.1.4",
  "pydantic>=2.7.0",
  "pypdf>=4.2.0",
  "pyyaml>=6.0.1",
  "tenacity>=8.2.3",
  "typer>=0.12.3",
  "rich>=13.7.1",
]

[project.scripts]
research-flow = "research_flow.cli:app"

[tool.setuptools]
packages = ["research_flow", "research_flow.providers", "research_flow.retrievers"]

